Inherit from new renewal reminder service base class
